Arreglos de Pago
Duración aproximada 4 horas

- Se tuvieron que corregir errores con el modelo de AP, principalmente se cambiaron los valores SqlMoney a Long.

-Se creó el metodo conexion Select que recibe los AP de una propiedad.

-Se modificaron el tipo de dato de algunos atributos del modelo de Recibos y Comprobates de Pago.

-Se agrego el atributo Metodo de pago para los Recibos por comprobante de pago

-Se agregó la lista de AP en el ViewModel de Propiedad Detail

-Se agregaron los recibos pendientes en la vista Detail de propiedad, asi como los AP.

-Se creó el store procedure SPS_AP_de_Propieadad que retorna los AP de una propiedad dado su numero de finca.

+++ Project_1/Models/Recibo/Recibo.cs
@@ -20,7 +20,7 @@ namespace Project_1.Models.Recibo
         public int id{ get; set; }
         public int idConceptoCobro{ get; set; }
         public String nombreConceptoCobro { get; set; }
-        public int monto{ get; set; }
+        public long monto{ get; set; }
         public DateTime fecha { get; set; }
         public DateTime fechaVencimiento { get; set; }
